Prompt Studio 하이라이트 코어 공통화
