+++ tests/test_aio_frontend.py
@@ -10,6 +10,7 @@ AIO_WHEEL_JS = ROOT / "web" / "js" / "aio" / "wheel.js"
 AIO_PRESETS_JS = ROOT / "web" / "js" / "aio" / "presets.js"
 AUTOCOMPLETE_JS = ROOT / "web" / "js" / "easyuse_anima_autocomplete.js"
 PROMPT_STUDIO_COMMON_JS = ROOT / "web" / "js" / "easyuse_anima_prompt_studio_common.js"
+PROMPT_STUDIO_HIGHLIGHT_CORE_JS = ROOT / "web" / "js" / "prompt_studio" / "highlight_core.js"
 
 
 class AIOFrontendSourceTests(unittest.TestCase):
@@ -641,17 +642,22 @@ class AIOFrontendSourceTests(unittest.TestCase):
         self.assertNotIn("easyuseAnimaDebugAutocomplete", source)
 
     def test_prompt_highlight_wildcards_accept_unicode_keys(self):
-        source = PROMPT_STUDIO_COMMON_JS.read_text(encoding="utf-8")
+        source = PROMPT_STUDIO_HIGHLIGHT_CORE_JS.read_text(encoding="utf-8")
+        common_source = PROMPT_STUDIO_COMMON_JS.read_text(encoding="utf-8")
 
         self.assertIn(r"const WILDCARD_HIGHLIGHT_RE = /(?:\d+#)?__[\p{L}\p{N}_.\-+/*\\]+?__/gu;", source)
         self.assertNotIn(r"const WILDCARD_HIGHLIGHT_RE = /(?:\d+#)?__[\w.\-+/*\\]+?__/g;", source)
 
         start = source.index("function renderHighlightedText")
-        end = source.index("\nfunction cssPixelNumber", start)
+        end = source.index("\n  return renderHighlightedText;", start)
         body = source[start:end]
 
-        self.assertLess(body.index("hasHighlightSyntax(body)"), body.index("const baseKey = normalize(tokenBase(body));"))
+        self.assertLess(
+            body.index("preferSyntaxBeforeToken && hasHighlightSyntax(body)"),
+            body.index("const baseKey = normalize(tokenBase(body));"),
+        )
         self.assertIn("html.push(syntaxHtml(body));", body)
+        self.assertIn("preferSyntaxBeforeToken: true", common_source)
 
 
 if __name__ == "__main__":
