+++ tools/check_frontend.ps1
@@ -29,6 +29,11 @@ try {
         }
     }
 
+    & node "tests\frontend_highlight_core_smoke.mjs"
+    if ($LASTEXITCODE -ne 0) {
+        throw "Frontend highlight core smoke failed with exit code $LASTEXITCODE."
+    }
+
     & npx --yes --package "typescript@$TypeScriptVersion" -- tsc -p jsconfig.json
     if ($LASTEXITCODE -ne 0) {
         throw "TypeScript check failed with exit code $LASTEXITCODE."
